fix MR.PARK use Genrative AI but no test
