+++ banana_api/serializers/model_file.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from ..models import ModelFile
+
+class ModelFileSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ModelFile
+        fields = '__all__'
