+++ banana_api/views/image_upload.py
@@ -1,14 +1,16 @@
 from rest_framework.views import APIView
-from ..utils.pagination import get_pagination_params
 from django.shortcuts import get_object_or_404
-from ..models import ModelFile
-from django.core.paginator import Paginator
-from ..serializers.model import ModelSerializer
 from django.http import JsonResponse
+from django.core.paginator import Paginator
+from rest_framework.response import Response
 
-class ModelList(APIView):
+from ..models import ImageUpload
+from ..serializers.image import ImageSerializer
+from ..utils.pagination import get_pagination_params
+
+class ImageList(APIView):
     """
-    GET /api/model/?page=<int>&size=<int>
+    GET /api/images/?page=<int>&size=<int>
 
     Query Parameters:
         - page (int, optional): Page number to retrieve. Default = 1
@@ -19,10 +21,11 @@ class ModelList(APIView):
     def get(self, request): 
         page, size = get_pagination_params(request)
         
-        ALLmodel = ModelFile.objects.all()
-        paginator = Paginator(ALLmodel, size)
+        ALLimage = ImageUpload.objects.all()
+        paginator = Paginator(ALLimage, size)
         page_obj = paginator.get_page(page)
-        serializer = ModelSerializer(page_obj, many=True)
+
+        serializer = ImageSerializer(page_obj, many=True)
 
         return JsonResponse({
             "count": paginator.count,
@@ -32,9 +35,9 @@ class ModelList(APIView):
             "results": serializer.data
         })
 
-class ModelDetailView(APIView):
+class ImageDetailView(APIView):
     """
-    GET /api/model/<pk>/
+    GET /api/images/<pk>/
 
     Path Parameters:
         - pk (int, required): The ID of the image to retrieve.
@@ -42,8 +45,6 @@ class ModelDetailView(APIView):
     Retrieves details of a specific image by ID.
     """
     def get(self, request, pk,):
-        image_obj = get_object_or_404(ModelFile, pk=pk)
-        serializer = ModelSerializer(image_obj)
-        return JsonResponse({
-            "results": serializer.data
-        })
+        image_obj = get_object_or_404(ImageUpload, pk=pk)
+        serializer = ImageSerializer(image_obj)
+        return Response(serializer.data)
